config/config.go
package config

import (
	"fmt"
	"log"
	"os"
	"path/filepath"

	"github.com/joho/godotenv"
	gormPostgres "gorm.io/driver/postgres"
	"gorm.io/gorm"
	"gorm.io/gorm/logger"
)

func LoadEnv(envFilePath string) {
	absPath, err := filepath.Abs(envFilePath)
	if err != nil {
		log.Fatalf("Error getting absolute path for %s: %v", envFilePath, err)
	}

	err = godotenv.Load(absPath)
	if err != nil {
		log.Fatalf("Error loading .env file at %s", absPath)
	}
}

func GetDatabase() *gorm.DB {
	dsn := fmt.Sprintf("host=%s port=%s user=%s password=%s dbname=%s sslmode=disable",
		os.Getenv("DB_HOST"), os.Getenv("DB_PORT"), os.Getenv("DB_USER"), os.Getenv("DB_PASSWORD"), os.Getenv("DB_NAME"))

	db, err := gorm.Open(gormPostgres.Open(dsn), &gorm.Config{
		Logger: logger.Default.LogMode(logger.Info),
	})
	if err != nil {
		log.Fatal("Failed to connect to database")
	}
	return db
}
